The full path is created as needed for the generated output.
The file save functionality needed to be changed to generate the saved parameters instead of expanding it to text, then saving it, to get this functionality to change.

A minor change of functionality was to only create the GenWriter object when a file name is specified. This allows the output destination to be specified dynamically.